pyproject.toml
[project]
name = "msw4fair"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.25",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
